fix: unify models used by tests to allow running them in a single command
