+++ test/webrtc_test.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-using_task_library "gstreamer"
+require_relative "models"
 
 WEBRTC_CONNECTED_STATES = {
     signaling_state: :GST_WEBRTC_SIGNALING_STATE_STABLE,
@@ -13,8 +13,9 @@ describe OroGen.gstreamer.WebRTCCommonTask do
     run_live
 
     before do
-        generator_m = OroGen.gstreamer.Task.specialize
-        generator_m.require_dynamic_service("image_source", as: "out")
+        generator_m = OroGen.gstreamer.Task.with_dynamic_service(
+            "image_source", as: "out"
+        )
 
         @sender_m = Syskit::Composition.new_submodel do
             add(generator_m, as: "generator")
@@ -337,68 +338,5 @@ describe OroGen.gstreamer.WebRTCCommonTask do
             .inject(&:+)
         mean_abs_sum = Float(abs_sum) / generator_frame.image.size
         mean_abs_sum < 10
-    end
-end
-
-module Services
-    data_service_type "ImageSource" do
-        output_port "image", ro_ptr("/base/samples/frame/Frame")
-    end
-
-    data_service_type "ImageSink" do
-        input_port "image", ro_ptr("/base/samples/frame/Frame")
-    end
-
-    data_service_type "DataChannel" do
-        input_port "in", "/iodrivers_base/RawPacket"
-        output_port "out", "/iodrivers_base/RawPacket"
-    end
-end
-
-Syskit.extend_model OroGen.gstreamer.WebRTCCommonTask do
-    argument :self_peer_id
-    argument :remote_peer_id, default: nil
-    argument :polite
-    argument :data_channels, default: []
-
-    dynamic_service Services::DataChannel, as: "data_channel" do
-        provides Services::DataChannel,
-                 as: name, "out" => "#{name}_out", "in" => "#{name}_in"
-    end
-
-    def update_properties
-        super
-
-        properties.signalling do |s|
-            s.self_peer_id = self_peer_id
-            s.remote_peer_id = remote_peer_id if remote_peer_id
-            s.polite = polite
-            s
-        end
-        properties.data_channels = data_channels
-    end
-end
-
-Syskit.extend_model OroGen.gstreamer.Task do
-    argument :pipeline
-    argument :inputs, default: []
-    argument :outputs, default: []
-    argument :channels, default: []
-
-    dynamic_service Services::ImageSink, as: "image_sink" do
-        provides Services::ImageSink, as: name, "image" => name
-    end
-
-    dynamic_service Services::ImageSource, as: "image_source" do
-        provides Services::ImageSource, as: name, "image" => name
-    end
-
-    def update_properties
-        super
-
-        properties.pipeline_initialization_timeout = Time.at(600)
-        properties.pipeline = pipeline
-        properties.inputs = inputs
-        properties.outputs = outputs
     end
 end
